setup/requirements.txt
psycopg2==2.9.9
mysql-connector-python==8.0.27
pandas==2.1.3
python-dotenv==1.0.0
google-cloud-storage==2.13.0
